fixed a bunch of compilation errors

+++ user-api/src/main/java/com/example/user_api/UserRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, int> {
     Optional<User> findByUserId(int userId);
 }
